Split out SingleKeyStamina class
