+++ osu.Game.Rulesets.Taiko/Difficulty/Skills/Stamina.cs
@@ -10,40 +10,6 @@ using osu.Game.Rulesets.Taiko.Objects;
 namespace osu.Game.Rulesets.Taiko.Difficulty.Skills
 {
     /// <summary>
-    /// Stamina of a single key, calculated based on repetition speed.
-    /// </summary>
-    public class SingleKeyStamina
-    {
-        private double? previousHitTime;
-
-        /// <summary>
-        /// Similar to <see cref="StrainDecaySkill.StrainValueOf(DifficultyHitObject)"/>
-        /// </summary>
-        public double StrainValueOf(DifficultyHitObject current)
-        {
-            if (previousHitTime == null)
-            {
-                previousHitTime = current.StartTime;
-                return 0;
-            }
-
-            double objectStrain = 0.5;
-            objectStrain += speedBonus(current.StartTime - previousHitTime.Value);
-            previousHitTime = current.StartTime;
-            return objectStrain;
-        }
-
-        /// <summary>
-        /// Applies a speed bonus dependent on the time since the last hit performed using this key.
-        /// </summary>
-        /// <param name="notePairDuration">The duration between the current and previous note hit using the same key.</param>
-        private double speedBonus(double notePairDuration)
-        {
-            return 175 / (notePairDuration + 100);
-        }
-    }
-
-    /// <summary>
     /// Calculates the stamina coefficient of taiko difficulty.
     /// </summary>
     /// <remarks>
